backend/requirements.txt
fastapi==0.115.12
uvicorn==0.33.0
starlette==0.44.0
pydantic==2.10.6
python-dotenv==1.0.1
typing-extensions==4.12.2
PyMySQL==1.1.0
cryptography>=39.0.0